Split into two versions
Decided to split this into two versions.
Downloader: which is the one-click download component ONLY
and
Companion: which will bring other utility such as notifications  as well
as one-click downloading.

But because some people might not like having an icon in the icon field
at all times just to download I split this into two versions