Forced re-cache of mangas after update
Put a system in place that will force an update to cached mangas after
an update to ensure that they are working correctly after an update in
case something was changed in the way they are stored or grabbed.

This should also come in handy if/when I bother to make an option to
check for updates automatically.